Tidy up namings and duplicate structs to version
