+++ runtime/src/serde_snapshot/common.rs
@@ -1,7 +1,5 @@
 use super::*;
-use crate::bank::RefBankFields;
 use std::collections::HashSet;
-use std::sync::RwLock;
 
 #[derive(Default, Clone, PartialEq, Debug, Deserialize, Serialize, AbiExample)]
 pub(crate) struct UnusedAccounts {
@@ -9,163 +7,3 @@ pub(crate) struct UnusedAccounts {
     unused2: HashSet<Pubkey>,
     unused3: HashMap<Pubkey, u64>,
 }
-
-// Deserializable version of Bank for snapshot format
-#[derive(Clone, Serialize, Deserialize, AbiExample)]
-pub(crate) struct DeserializableBankFields {
-    pub(crate) blockhash_queue: BlockhashQueue,
-    pub(crate) ancestors: Ancestors,
-    pub(crate) hash: Hash,
-    pub(crate) parent_hash: Hash,
-    pub(crate) parent_slot: Slot,
-    pub(crate) hard_forks: HardForks,
-    pub(crate) transaction_count: u64,
-    pub(crate) tick_height: u64,
-    pub(crate) signature_count: u64,
-    pub(crate) capitalization: u64,
-    pub(crate) max_tick_height: u64,
-    pub(crate) hashes_per_tick: Option<u64>,
-    pub(crate) ticks_per_slot: u64,
-    pub(crate) ns_per_slot: u128,
-    pub(crate) genesis_creation_time: UnixTimestamp,
-    pub(crate) slots_per_year: f64,
-    pub(crate) unused: u64,
-    pub(crate) slot: Slot,
-    pub(crate) epoch: Epoch,
-    pub(crate) block_height: u64,
-    pub(crate) collector_id: Pubkey,
-    pub(crate) collector_fees: u64,
-    pub(crate) fee_calculator: FeeCalculator,
-    pub(crate) fee_rate_governor: FeeRateGovernor,
-    pub(crate) collected_rent: u64,
-    pub(crate) rent_collector: RentCollector,
-    pub(crate) epoch_schedule: EpochSchedule,
-    pub(crate) inflation: Inflation,
-    pub(crate) stakes: Stakes,
-    pub(crate) unused_accounts: UnusedAccounts,
-    pub(crate) epoch_stakes: HashMap<Epoch, EpochStakes>,
-    pub(crate) is_delta: bool,
-    pub(crate) message_processor: MessageProcessor,
-}
-
-impl Into<BankFields> for DeserializableBankFields {
-    fn into(self) -> BankFields {
-        BankFields {
-            blockhash_queue: self.blockhash_queue,
-            ancestors: self.ancestors,
-            hash: self.hash,
-            parent_hash: self.parent_hash,
-            parent_slot: self.parent_slot,
-            hard_forks: self.hard_forks,
-            transaction_count: self.transaction_count,
-            tick_height: self.tick_height,
-            signature_count: self.signature_count,
-            capitalization: self.capitalization,
-            max_tick_height: self.max_tick_height,
-            hashes_per_tick: self.hashes_per_tick,
-            ticks_per_slot: self.ticks_per_slot,
-            ns_per_slot: self.ns_per_slot,
-            genesis_creation_time: self.genesis_creation_time,
-            slots_per_year: self.slots_per_year,
-            unused: self.unused,
-            slot: self.slot,
-            epoch: self.epoch,
-            block_height: self.block_height,
-            collector_id: self.collector_id,
-            collector_fees: self.collector_fees,
-            fee_calculator: self.fee_calculator,
-            fee_rate_governor: self.fee_rate_governor,
-            collected_rent: self.collected_rent,
-            rent_collector: self.rent_collector,
-            epoch_schedule: self.epoch_schedule,
-            inflation: self.inflation,
-            stakes: self.stakes,
-            epoch_stakes: self.epoch_stakes,
-            is_delta: self.is_delta,
-        }
-    }
-}
-
-// Serializable version of Bank for snapshot format
-#[derive(Serialize)]
-pub(crate) struct SerializableBankFields<'a> {
-    pub(crate) blockhash_queue: &'a RwLock<BlockhashQueue>,
-    pub(crate) ancestors: &'a Ancestors,
-    pub(crate) hash: Hash,
-    pub(crate) parent_hash: Hash,
-    pub(crate) parent_slot: Slot,
-    pub(crate) hard_forks: &'a RwLock<HardForks>,
-    pub(crate) transaction_count: u64,
-    pub(crate) tick_height: u64,
-    pub(crate) signature_count: u64,
-    pub(crate) capitalization: u64,
-    pub(crate) max_tick_height: u64,
-    pub(crate) hashes_per_tick: Option<u64>,
-    pub(crate) ticks_per_slot: u64,
-    pub(crate) ns_per_slot: u128,
-    pub(crate) genesis_creation_time: UnixTimestamp,
-    pub(crate) slots_per_year: f64,
-    pub(crate) unused: u64,
-    pub(crate) slot: Slot,
-    pub(crate) epoch: Epoch,
-    pub(crate) block_height: u64,
-    pub(crate) collector_id: Pubkey,
-    pub(crate) collector_fees: u64,
-    pub(crate) fee_calculator: FeeCalculator,
-    pub(crate) fee_rate_governor: FeeRateGovernor,
-    pub(crate) collected_rent: u64,
-    pub(crate) rent_collector: RentCollector,
-    pub(crate) epoch_schedule: EpochSchedule,
-    pub(crate) inflation: Inflation,
-    pub(crate) stakes: &'a RwLock<Stakes>,
-    pub(crate) unused_accounts: UnusedAccounts,
-    pub(crate) epoch_stakes: &'a HashMap<Epoch, EpochStakes>,
-    pub(crate) is_delta: bool,
-    pub(crate) message_processor: MessageProcessor,
-}
-
-impl<'a> From<RefBankFields<'a>> for SerializableBankFields<'a> {
-    fn from(rhs: RefBankFields<'a>) -> Self {
-        fn new<T: Default>() -> T {
-            T::default()
-        }
-        Self {
-            blockhash_queue: rhs.blockhash_queue,
-            ancestors: rhs.ancestors,
-            hash: rhs.hash,
-            parent_hash: rhs.parent_hash,
-            parent_slot: rhs.parent_slot,
-            hard_forks: rhs.hard_forks,
-            transaction_count: rhs.transaction_count,
-            tick_height: rhs.tick_height,
-            signature_count: rhs.signature_count,
-            capitalization: rhs.capitalization,
-            max_tick_height: rhs.max_tick_height,
-            hashes_per_tick: rhs.hashes_per_tick,
-            ticks_per_slot: rhs.ticks_per_slot,
-            ns_per_slot: rhs.ns_per_slot,
-            genesis_creation_time: rhs.genesis_creation_time,
-            slots_per_year: rhs.slots_per_year,
-            unused: rhs.unused,
-            slot: rhs.slot,
-            epoch: rhs.epoch,
-            block_height: rhs.block_height,
-            collector_id: rhs.collector_id,
-            collector_fees: rhs.collector_fees,
-            fee_calculator: rhs.fee_calculator,
-            fee_rate_governor: rhs.fee_rate_governor,
-            collected_rent: rhs.collected_rent,
-            rent_collector: rhs.rent_collector,
-            epoch_schedule: rhs.epoch_schedule,
-            inflation: rhs.inflation,
-            stakes: rhs.stakes,
-            unused_accounts: new(),
-            epoch_stakes: rhs.epoch_stakes,
-            is_delta: rhs.is_delta,
-            message_processor: new(),
-        }
-    }
-}
-
-#[cfg(RUSTC_WITH_SPECIALIZATION)]
-impl<'a> IgnoreAsHelper for SerializableBankFields<'a> {}
